+++ src/update.cpp
@@ -5,8 +5,17 @@
 #include <QJsonDocument>
 #include <QProcess>
 #include <QStandardPaths>
+#include <QCryptographicHash>
 #include "logger.hpp"
 extern Logger *lg;
+
+namespace {
+QByteArray sha256file(QFile &file) {
+  QCryptographicHash hash(QCryptographicHash::Sha256);
+  if (!hash.addData(&file)) return {};
+  return hash.result().toHex();
+}
+} // namespace
 
 UpdateManager::UpdateManager(QObject *parent)
     : QObject(parent), m_netman(new QNetworkAccessManager(this)) {
@@ -16,8 +25,14 @@ UpdateManager::UpdateManager(QObject *parent)
 // https://github.com/ilpenSE/simpleclicker/releases/download/v1.1.0-beta/SimpleClicker.exe
 // https://github.com/ilpenSE/simpleclicker/releases/download/v1.1.0-beta/gnu-linux-x86-64.tar.gz
 void UpdateManager::downloadAndInstall(Version version) {
-  QString url = QString("https://github.com/ilpenSE/simpleclicker/releases/download/v%1/SimpleClicker.exe")
-                        .arg(version.toQString());
+  QString url = QString("https://github.com/ilpenSE/simpleclicker/releases/download/v%1/%2")
+                        .arg(version.toQString())
+#ifdef _WIN32
+                        .arg("SimpleClicker.exe");
+#else
+                        .arg("gnu-linux-x86-64.tar.gz");
+#endif
+
   lg->info("Downloading: {}", url);
 
   QNetworkRequest request{QUrl(url)};
@@ -32,6 +47,8 @@ void UpdateManager::downloadAndInstall(Version version) {
                           .filePath("SimpleClicker_setup.tar.gz");
 #endif
 
+  // TODO: Fetch latest release binary's checksum first and compare existing file (if exists)
+  // if they're not equal only then install it
   auto *file = new QFile(tempPath, this);
   if (!file->open(QIODevice::WriteOnly)) {
     emit downloadFailed("Could not open temporary file: " + tempPath);
@@ -65,51 +82,66 @@ void UpdateManager::downloadAndInstall(Version version) {
     }
     lg->info("Downloaded file to {}", tempPath);
 
+    install(tempPath);
+  });
+}
+
+void UpdateManager::install(const QString &scriptPath) {
 #ifdef _WIN32
-    qint64 pid = 0;
-    bool started = QProcess::startDetached(tempPath, {"/VERYSILENT", "/NORESTART"}, QString(), &pid);
-    if (!started) {
-      emit downloadFailed(QString("Could not start installer with pid: ").arg(pid));
-      return;
-    }
-#else
-    QProcess process;
+  qint64 pid = 0;
+  bool started = QProcess::startDetached(scriptPath, {"/VERYSILENT", "/NORESTART"}, QString(), &pid);
+  if (!started) {
+    emit downloadFailed(QString("Could not start installer with pid: %1").arg(pid));
+    return;
+  }
 
-    // Unpack the tarball downloaded
-    process.start("tar", {"-xzf", tempPath, "-C", "/tmp/"});
-    if (!process.waitForStarted()) {
-      emit downloadFailed("Could not create tar process");
-      return;
-    }
-    if (!process.waitForFinished(-1)) {
-      emit downloadFailed("Tar process exited abnormally: " + process.errorString());
-      return;
-    }
+#else // Linux
+  QProcess process;
 
-    // Start install script
-    process.start("pkexec", {"/tmp/simpleclicker/install.sh", "/tmp/simpleclicker/"});
-    if (!process.waitForStarted()) {
-      emit downloadFailed("Could not create install script process");
-      return;
-    }
-    if (!process.waitForFinished(-1)) {
-      emit downloadFailed("Install script exited abnormally: " + process.errorString());
-      return;
-    }
-    lg->info("Install script says (stdout):\n{}", QString(process.readAllStandardOutput()));
-    lg->info("Install script says (stderr):\n{}", QString(process.readAllStandardError()));
+  // Unpack the tarball downloaded
+  process.start("tar", {"-xzf", scriptPath, "-C", "/tmp/"});
+  if (!process.waitForStarted()) {
+    emit downloadFailed("Could not create tar process");
+    return;
+  }
+  if (!process.waitForFinished(-1)) {
+    emit downloadFailed("Tar process exited abnormally: " + process.errorString());
+    return;
+  }
 
-    // Open upgraded program
-    qint64 pid;
-    process.setProgram("/usr/local/bin/SimpleClicker");
-    if (!process.startDetached(&pid)) {
-      lg->error("Could not open new program");
-      return;
-    }
+  // Start install script
+  process.start("pkexec", {"/tmp/simpleclicker/install.sh", "/tmp/simpleclicker/"});
+  if (!process.waitForStarted()) {
+    emit downloadFailed("Could not create install script process");
+    return;
+  }
+  if (!process.waitForFinished(-1)) {
+    emit downloadFailed("Install script exited abnormally: " + process.errorString());
+    return;
+  }
+  lg->info("Install script's stdout:\n{}", QString(process.readAllStandardOutput()));
+  lg->info("Install script's stderr:\n{}", QString(process.readAllStandardError()));
+
+  const int exitCode = process.exitCode();
+  // User doesn't authorized, graceful fail
+  if (exitCode == 127 || exitCode == 126) return;
+
+  // Real error
+  if (exitCode != 0) {
+    emit downloadFailed(QString("Install script failed with exit code %1").arg(exitCode));
+    return;
+  }
+
+  // Open upgraded program
+  qint64 pid;
+  process.setProgram("/usr/local/bin/SimpleClicker");
+  if (!process.startDetached(&pid)) {
+    lg->error("Could not open new program");
+    return;
+  }
 #endif
 
-    qApp->quit();
-  });
+  qApp->quit();
 }
 
 void UpdateManager::checkForUpdates() {